Fix the Model the Role

+++ resources/views/admin/sale/index.blade.php
@@ -29,7 +29,9 @@
 
                         <div class="d-flex justify-content-between">
                             <h4 class="card-title">Ventas
-                                <a href="{{ route('sales.create') }}" class="btn btn-primary">Crear Nueva</a>
+                                @can('sales.create')
+                                    <a href="{{ route('sales.create') }}" class="btn btn-primary">Crear Nueva</a>
+                                @endcan
                             </h4>
                             <div class="btn-group">
                                 <a data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
@@ -47,7 +49,9 @@
                                         <th>Codigo</th>
                                         <th>Cliente</th>
                                         <th>Fecha de Venta</th>
-                                        <th>Estado</th>
+                                        @can('sale.change_status')
+                                            <th>Estado</th>
+                                        @endcan
                                         <th>Impuesto</th>
                                         <th>Total</th>
                                         <th style="width: 5px;">Acciones</th>
@@ -60,23 +64,29 @@
                                             <td>{{ $sale->name }}</td>
                                             <td>{{ $sale->sale_date }}</td>
                                             @if ($sale->status === 'VALID')
+                                            @can('sale.change_status')
                                                 <td style="width: 50px;">
                                                     <a class="jsgrid-button btn btn-success" href="{{ route('sale.change_status', $sale) }}"> Valida<i class="fas fa-check"></i></a>
                                                 </td>
+                                            @endcan
                                             @else
+                                            @can('sale.change_status')
                                                 <td>
                                                     <a class="jsgrid-button btn btn-danger" href="{{ route('sale.change_status', $sale) }}"> Cancelada <i class="fas fa-times"></i></a>
                                                 </td>
+                                            @endcan
                                             @endif
                                             <td>{{ $sale->impuesto }}</td>
                                             <td>{{ $sale->total }}</td>
                                             <td style="width: 100px;">
-                                                {!! Form::open(['route' => ['sales.destroy', $sale], 'method' => 'DELETE']) !!}
-                                                <a class="jsgrid-button jsgrid-edit-button" href="{{ route('sales.show', $sale->id) }}" title="Ver Detalle"><i class="fas fa-eye"></i></a>
-                                                <a class="jsgrid-button jsgrid-edit-button" href="{{ route('sales.pdf', $sale->id) }}" title="Reporte PDF"><i class="fas fa-file-pdf"></i></a>
-                                                <a class="jsgrid-button jsgrid-edit-button" href="{{ route('sales.print',$sale->id) }}" id="print" title="Imprimir"><i class="fas fa-print"></i></a>
-                                                <button class="jsgrid-button jsgrid-delete-button unstyled-button" type="submit" title="Anular Compra"><i class="far fa-trash-alt"></i></button>
-                                                {!! Form::close() !!}
+                                                @can('sales.destroy','sales.show','sales.pdf','sales.print')
+                                                    {!! Form::open(['route' => ['sales.destroy', $sale], 'method' => 'DELETE']) !!}
+                                                    <a class="jsgrid-button jsgrid-edit-button" href="{{ route('sales.show', $sale->id) }}" title="Ver Detalle"><i class="fas fa-eye"></i></a>
+                                                    <a class="jsgrid-button jsgrid-edit-button" href="{{ route('sales.pdf', $sale->id) }}" title="Reporte PDF"><i class="fas fa-file-pdf"></i></a>
+                                                    <a class="jsgrid-button jsgrid-edit-button" href="{{ route('sales.print',$sale->id) }}" id="print" title="Imprimir"><i class="fas fa-print"></i></a>
+                                                    <button class="jsgrid-button jsgrid-delete-button unstyled-button" type="submit" title="Anular Compra"><i class="far fa-trash-alt"></i></button>
+                                                    {!! Form::close() !!}
+                                                @endcan
                                             </td>
                                         </tr>
                                     @endforeach
